+++ views/default/resources/tidypics/photos/album/add.php
@@ -6,7 +6,7 @@
  * @license http://www.gnu.org/licenses/gpl-2.0.html GNU General Public License v2
  */
 
-$guid = elgg_extract('guid', $vars);
+$guid = (int) elgg_extract('guid', $vars);
 if (!$guid) {
 	$guid = elgg_get_logged_in_user_guid();
 }
